Update Calendar.md
Regarding remarks: 

•	We do not want to refer to other ERPs. - Reference to other ERPs removed.
•	We do not want negative connotation in any aspect. - Negative connotation changed.
•	We just affirm our advantages. - Emphasized on multiple legal entities support, user-friendly interface, easy navigation, filtering combinations, and other features.

Regarding remarks: 

What are these? If this is user-defined, this should be specified. - Document types are now described with more clarity. Description of what is system-based and predefined, and what is user-configurable or vendor-configurable. 

Regarding remarks: 

This is not automatic. It is the basic type which the user chooses when creating new Activity. - Fixed the "automatic" statement. Changed to user choice between system types.