Update and rename Concurrent_sessions_based_licensing.md to concurrent-sessions-based-licensing.md
Regarding remarks: 

The note for the servers is unnecessary. Just specify the user actions. - The note for the servers is removed. Used "@@name (ERP.net) instance" instead. 

Regarding remarks: 

It's not a "database", it's an "instance" or "ERP instance". - "Database" removed. Used "ERP.net instance for this user" instead. 

Regarding remarks: 

Avoid talking about "us", unless absolutely necessary. Concentrate on the user. - Stated in a different way, addressing the user/client instead of "our clients". 

Regarding remarks: 

It is not of our business to talk about "your staff members" in the licensing section. State it in a different way. - Addressing as "your staff members" was removed. Stated in a non-client's staff members reference way. 

Regarding remarks: 

Change the file name to be in kebab-case and not snake_case. - The file was named in kebab-case style, in accordance with other files in this directory.